# ecommerce.orders.manager/docs/swagger.yaml
basePath: /
definitions:
...
  ecommerce_orders_manager_internal_models.TokenResponse:
    properties:
      access_token:
        type: string
      token_type:
        type: string
    type: object
...
host: localhost:8082
info:
  contact: {}
  description: Orders Manager service for e-commerce microservices
  title: Orders Manager API
  version: "1.0"
paths:
  /auth/token:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: OAuth2 password flow - exchange username/password for access token
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ecommerce_orders_manager_internal_models.TokenResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get access token
      tags:
      - Authentication
...
securityDefinitions:
  OAuth2PasswordBearer:
    flow: password
    tokenUrl: /auth/token
    type: oauth2
swagger: "2.0"